+++ src/features/events/actions/endEventAction.ts
@@ -70,6 +70,11 @@ export async function endEventAction(event_id: string): Promise<ActionResponse<v
     ended_at: new Date(),
   });
 
+  //Mark the event as ended on each participant.
+  await db(tablenames.event_attendance).where({ event_instance_id: event_id }).update({
+    event_ended: true,
+  });
+
   //global.io.to('user:' + session.user.id).emit('event_ended');
   global.io.to('event:' + event_id).emit('event:end', { eventId: event_id });
   return { success: true };