+++ src/features/events/providers/EventProvider.tsx
@@ -10,6 +10,8 @@ import { useReloadData } from '@/hooks/useReloadData';
 import { useRouter } from 'next/navigation';
 import { useGeolocationContext } from '@/features/geolocation/providers/GeolocationProvider';
 import { useDistance } from '../../distance/hooks/useDistance';
+import { endEventAction } from '../actions/endEventAction';
+import { useUserAttendanceContext } from '@/features/attendance/providers/UserAttendanceProvider';
 
 type EventProviderProps = React.PropsWithChildren & {
   initialEvent: TEvent;
@@ -17,11 +19,13 @@ type EventProviderProps = React.PropsWithChildren & {
 
 const [EventContext, useEventContext] = createContextWithUseHook<{
   event: EventProviderProps['initialEvent'];
+  end: () => Promise<void>;
   hasEnded: boolean;
   interestCount: number;
 }>('useEventContext can only be used within the scope of an EventContext!');
 
 export function EventProvider({ children, initialEvent }: EventProviderProps) {
+  const attendance = useUserAttendanceContext();
   const [event, setEvent] = useState(initialEvent);
   const hasEnded = event.ended_at !== null;
   const interestCount = event.interested_count;
@@ -34,6 +38,11 @@ export function EventProvider({ children, initialEvent }: EventProviderProps) {
     300
   );
 
+  const end = async () => {
+    await endEventAction(event.id);
+    await attendance.leave(event.id);
+  };
+
   useEventSocket({
     eventId: event.id,
     onEnd: () => reloadEvent(),
@@ -41,7 +50,7 @@ export function EventProvider({ children, initialEvent }: EventProviderProps) {
   });
 
   return (
-    <EventContext.Provider value={{ event, hasEnded, interestCount }}>
+    <EventContext.Provider value={{ event, hasEnded, interestCount, end }}>
       {children}
     </EventContext.Provider>
   );