+++ src/middleware.ts
@@ -21,7 +21,7 @@ export default async function middleware(req: NextRequestWithAuth) {
       return NextResponse.redirect(newUrl);
     }
 
-    if (token.attended_event_id) {
+    /*if (token.attended_event_id) {
       if (
         url.pathname === '/' ||
         url.pathname.startsWith('/app/event/create') ||
@@ -31,7 +31,7 @@ export default async function middleware(req: NextRequestWithAuth) {
         newUrl.pathname = '/app/event/' + token.attended_event_id;
         return NextResponse.redirect(newUrl);
       }
-    } else if (
+    } else*/ if (
       url.pathname === '/' ||
       url.pathname.includes('/login') ||
       url.pathname.includes('/register') ||